move pagemeta from format => cmdctx
