+++ .gitignore
@@ -21,3 +21,4 @@ internal-docs/
 .claude/settings.local.md
 .claude/
 .cursor/
+devhome/
